Include unauthorized parents on startup.
For the navigation to work properly, we need to make sure that
the parents are included on startup.
Otherwise no tree can be built and the navigation would not work.
There might be the problem that the current user is not authorized
to access a parent (View permission).
For beeing consistent with other Plone navigations such as the
navigation portlet or breadcrumbs we want to render the parents
anyway and let the user hit an "Anauthorized"-error when the user
clicks.
This has the advantage that we can build the navigation tree properly.

This method prepends the missing parents by doing an unrestricted
catalog query.