make navigation view public.
The JavaScript may fetch children now and then, sometimes even
on objects on which the current user has no View permission.
In order for the JavaScript to work properly, this should not fail
with an unauthorized error but return possible children.

This might be a situation where the unauthorized object was already
loaded as parent and the JavaScript is trying to fetch its children,
which may even be published / visible.