protect the startup endpoint since it is now public.
For security it is essential that the `startup` endpoint does
not work on a context where the current user has no `View` permission.
This is not protected by ZCML since the `children` endpoint needs
to be public.

The `startup` endpoint may provide infos of unpublished content (such
as the title). This is necessary for the navigation to work and follows
Plone principal but it should not be possible to misuse the endpoint
for getting those informations on _any_ content.