Fix crash when parent is excluded from navigation.
When the navigation is initialized on a context of which the parent is
excluded from navigation, it crashed, because it could not render the
"go up" link.

In this situation we show the normally not shown parent since the start
point is within this parent. This is consistent with the desktop
behavior (e.g. breadcrumbs, navigation portlet).