Install ftw.gopip for better performance.
ftw.mobile does large catalog queries ordered by the
getObjPositionInParent, which is actually very slow because it loads the
objects.

ftw.gopip improves the getObjPositionInParent index read performance by
replacing it with a FieldIndex which actually stores the position so
that the objects must no longer be woken up for sorting by position.